requirements.txt
motor
Kurigram
pyromod
TgCrypto
python-dotenv
telethon
rich